feat: add training scripts
- make training more streamlined
