+++ sitecustomize.py
@@ -1,3 +1,10 @@
-import coverage
+import os
 
-coverage.process_startup()
+# Only import and start coverage when COVERAGE_PROCESS_START is set
+if os.environ.get("COVERAGE_PROCESS_START"):
+    try:
+        import coverage
+
+        coverage.process_startup()
+    except ImportError:
+        pass
